Fixes to use the container name as the queue / topic name, password / protect the connection string, and do not display the connection string in the account name.
NB, for some consumers the change to use the container name rather than the account name for sending may cause backward compatibility issues. It would be better to use the container name as the queue name so that multiple streams can be used for the same Export Target, e.g. as done with SQL